Use unique name for luk-org-paste-image
When writing the png in clipboard_to_file.py, instead of always using
"paste.png" as the initial filename (which can then be edited
interactively), check if the name is available, and if it isn't, find
the first available name on the style "paste-<num>.png".

This avoids overwriting paste.png in the org attachment folder for the
section if the user did not rename the previous paste.