Move default font initialization to "luk-font.el" and add checks
Move the `set-face-attribute` for the `default`-font to "luk-font.el"

Add checks for font existence before setting it. Add Consolas as an
alternative if DejaVu Sans Mono is missing.

Also check for font existence before all calls to `set-fontset-font`
in "luk-font.el", to avoid specifying a missing font for unicode code
points.