Prevent rank-change in capture
It seems bad to be able to change the rank of an org-heading while in
a capture buffer, so naively rebind M-<left>, M-<right> to show
warnings.

Note: It would be better to do something smarter, like preventing the
topmost node from being demoted or removed at all while allowing
reordering child nodes, but that's more difficult, especially since
the ongoing capture can be affected by changes in the target buffer
outside of the capture window.